fix: file name and close scanner
